maintain: split handlers into separate files
