+++ internal/server/handlers.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"regexp"
-	"strings"
 	"time"
 
 	"github.com/getkin/kin-openapi/openapi3"
@@ -17,10 +15,8 @@ import (
 	"github.com/infrahq/infra/internal/access"
 	"github.com/infrahq/infra/internal/logging"
 	"github.com/infrahq/infra/internal/server/authn"
-	"github.com/infrahq/infra/internal/server/data"
 	"github.com/infrahq/infra/internal/server/models"
 	"github.com/infrahq/infra/internal/server/providers"
-	"github.com/infrahq/infra/uid"
 )
 
 type API struct {
@@ -28,319 +24,6 @@ type API struct {
 	server     *Server
 	migrations []apiMigration
 	openAPIDoc openapi3.T
-}
-
-func (a *API) ListUsers(c *gin.Context, r *api.ListUsersRequest) (*api.ListResponse[api.User], error) {
-	p := models.RequestToPagination(r.PaginationRequest)
-	users, err := access.ListIdentities(c, r.Name, r.Group, r.IDs, &p)
-	if err != nil {
-		return nil, err
-	}
-
-	result := api.NewListResponse(users, models.PaginationToResponse(p), func(identity models.Identity) api.User {
-		return *identity.ToAPI()
-	})
-
-	return result, nil
-}
-
-func (a *API) GetUser(c *gin.Context, r *api.GetUserRequest) (*api.User, error) {
-	if r.ID.IsSelf {
-		iden := access.AuthenticatedIdentity(c)
-		if iden == nil {
-			return nil, fmt.Errorf("%w: no user is logged in", internal.ErrUnauthorized)
-		}
-		r.ID.ID = iden.ID
-	}
-	identity, err := access.GetIdentity(c, r.ID.ID)
-	if err != nil {
-		return nil, err
-	}
-
-	return identity.ToAPI(), nil
-}
-
-// CreateUser creates a user with the Infra provider
-func (a *API) CreateUser(c *gin.Context, r *api.CreateUserRequest) (*api.CreateUserResponse, error) {
-	user := &models.Identity{Name: r.Name}
-	infraProvider := access.InfraProvider(c)
-
-	// infra identity creation should be attempted even if an identity is already known
-	identities, err := access.ListIdentities(c, user.Name, 0, nil, &models.Pagination{Limit: 2})
-	if err != nil {
-		return nil, fmt.Errorf("list identities: %w", err)
-	}
-
-	switch len(identities) {
-	case 0:
-		if err := access.CreateIdentity(c, user); err != nil {
-			return nil, fmt.Errorf("create identity: %w", err)
-		}
-	case 1:
-		user.ID = identities[0].ID
-	default:
-		logging.Errorf("Multiple identites match name %q. DB is missing unique index on user names", r.Name)
-		return nil, fmt.Errorf("multiple identities match specified name") // should not happen
-	}
-
-	resp := &api.CreateUserResponse{
-		ID:   user.ID,
-		Name: user.Name,
-	}
-
-	_, err = access.CreateProviderUser(c, infraProvider, user)
-	if err != nil {
-		return nil, fmt.Errorf("creating provider user: %w", err)
-	}
-
-	// Always create a temporary password for infra users.
-	tmpPassword, err := access.CreateCredential(c, *user)
-	if err != nil {
-		return nil, fmt.Errorf("create credential: %w", err)
-	}
-
-	resp.OneTimePassword = tmpPassword
-
-	return resp, nil
-}
-
-func (a *API) UpdateUser(c *gin.Context, r *api.UpdateUserRequest) (*api.User, error) {
-	// right now this endpoint can only update a user's credentials, so get the user identity
-	identity, err := access.GetIdentity(c, r.ID)
-	if err != nil {
-		return nil, err
-	}
-
-	err = access.UpdateCredential(c, identity, r.Password)
-	if err != nil {
-		return nil, err
-	}
-
-	// if the user is an admin, we could be required to create the infra user, so create the provider_user if it's missing.
-	_, _ = access.CreateProviderUser(c, access.InfraProvider(c), identity)
-
-	return identity.ToAPI(), nil
-}
-
-func (a *API) DeleteUser(c *gin.Context, r *api.Resource) (*api.EmptyResponse, error) {
-	return nil, access.DeleteIdentity(c, r.ID)
-}
-
-// TODO: remove after deprecation period
-func (a *API) deprecatedListUserGroups(c *gin.Context, r *api.Resource) (*api.ListResponse[api.Group], error) {
-	return a.ListGroups(c, &api.ListGroupsRequest{UserID: r.ID})
-}
-
-func (a *API) ListGroups(c *gin.Context, r *api.ListGroupsRequest) (*api.ListResponse[api.Group], error) {
-	p := models.RequestToPagination(r.PaginationRequest)
-	groups, err := access.ListGroups(c, r.Name, r.UserID, &p)
-	if err != nil {
-		return nil, err
-	}
-
-	result := api.NewListResponse(groups, models.PaginationToResponse(p), func(group models.Group) api.Group {
-		return *group.ToAPI()
-	})
-
-	return result, nil
-}
-
-func (a *API) GetGroup(c *gin.Context, r *api.Resource) (*api.Group, error) {
-	group, err := access.GetGroup(c, r.ID)
-	if err != nil {
-		return nil, err
-	}
-
-	return group.ToAPI(), nil
-}
-
-func (a *API) CreateGroup(c *gin.Context, r *api.CreateGroupRequest) (*api.Group, error) {
-	group := &models.Group{
-		Name: r.Name,
-	}
-
-	authIdent := access.AuthenticatedIdentity(c)
-	if authIdent != nil {
-		group.CreatedBy = authIdent.ID
-	}
-
-	err := access.CreateGroup(c, group)
-	if err != nil {
-		return nil, err
-	}
-
-	return group.ToAPI(), nil
-}
-
-func (a *API) DeleteGroup(c *gin.Context, r *api.Resource) (*api.EmptyResponse, error) {
-	return nil, access.DeleteGroup(c, r.ID)
-}
-
-func (a *API) UpdateUsersInGroup(c *gin.Context, r *api.UpdateUsersInGroupRequest) (*api.EmptyResponse, error) {
-	return nil, access.UpdateUsersInGroup(c, r.GroupID, r.UserIDsToAdd, r.UserIDsToRemove)
-}
-
-// caution: this endpoint is unauthenticated, do not return sensitive info
-func (a *API) ListProviders(c *gin.Context, r *api.ListProvidersRequest) (*api.ListResponse[api.Provider], error) {
-	exclude := []models.ProviderKind{models.ProviderKindInfra}
-	p := models.RequestToPagination(r.PaginationRequest)
-	providers, err := access.ListProviders(c, r.Name, exclude, &p)
-	if err != nil {
-		return nil, err
-	}
-
-	result := api.NewListResponse(providers, models.PaginationToResponse(p), func(provider models.Provider) api.Provider {
-		return *provider.ToAPI()
-	})
-
-	return result, nil
-}
-
-// caution: this endpoint is unauthenticated, do not return sensitive info
-func (a *API) GetProvider(c *gin.Context, r *api.Resource) (*api.Provider, error) {
-	provider, err := access.GetProvider(c, r.ID)
-	if err != nil {
-		return nil, err
-	}
-
-	return provider.ToAPI(), nil
-}
-
-var (
-	dashAdminRemover = regexp.MustCompile(`(.*)\-admin(\.okta\.com)`)
-	protocolRemover  = regexp.MustCompile(`http[s]?://`)
-)
-
-func cleanupURL(url string) string {
-	url = strings.TrimSpace(url)
-	url = dashAdminRemover.ReplaceAllString(url, "$1$2")
-	url = protocolRemover.ReplaceAllString(url, "")
-
-	return url
-}
-
-func (a *API) CreateProvider(c *gin.Context, r *api.CreateProviderRequest) (*api.Provider, error) {
-	provider := &models.Provider{
-		Name:         r.Name,
-		URL:          cleanupURL(r.URL),
-		ClientID:     r.ClientID,
-		ClientSecret: models.EncryptedAtRest(r.ClientSecret),
-	}
-
-	kind, err := models.ParseProviderKind(r.Kind)
-	if err != nil {
-		return nil, err
-	}
-	provider.Kind = kind
-
-	if err := a.setProviderInfoFromServer(c, provider); err != nil {
-		return nil, err
-	}
-
-	if err := access.CreateProvider(c, provider); err != nil {
-		return nil, err
-	}
-
-	return provider.ToAPI(), nil
-}
-
-func (a *API) UpdateProvider(c *gin.Context, r *api.UpdateProviderRequest) (*api.Provider, error) {
-	provider := &models.Provider{
-		Model: models.Model{
-			ID: r.ID,
-		},
-		Name:         r.Name,
-		URL:          cleanupURL(r.URL),
-		ClientID:     r.ClientID,
-		ClientSecret: models.EncryptedAtRest(r.ClientSecret),
-	}
-
-	kind, err := models.ParseProviderKind(r.Kind)
-	if err != nil {
-		return nil, err
-	}
-	provider.Kind = kind
-
-	if err := a.setProviderInfoFromServer(c, provider); err != nil {
-		return nil, err
-	}
-
-	if err := access.SaveProvider(c, provider); err != nil {
-		return nil, err
-	}
-
-	return provider.ToAPI(), nil
-}
-
-func (a *API) DeleteProvider(c *gin.Context, r *api.Resource) (*api.EmptyResponse, error) {
-	return nil, access.DeleteProvider(c, r.ID)
-}
-
-func (a *API) ListDestinations(c *gin.Context, r *api.ListDestinationsRequest) (*api.ListResponse[api.Destination], error) {
-	p := models.RequestToPagination(r.PaginationRequest)
-	destinations, err := access.ListDestinations(c, r.UniqueID, r.Name, &p)
-	if err != nil {
-		return nil, err
-	}
-
-	result := api.NewListResponse(destinations, models.PaginationToResponse(p), func(destination models.Destination) api.Destination {
-		return *destination.ToAPI()
-	})
-
-	return result, nil
-}
-
-func (a *API) GetDestination(c *gin.Context, r *api.Resource) (*api.Destination, error) {
-	destination, err := access.GetDestination(c, r.ID)
-	if err != nil {
-		return nil, err
-	}
-
-	return destination.ToAPI(), nil
-}
-
-func (a *API) CreateDestination(c *gin.Context, r *api.CreateDestinationRequest) (*api.Destination, error) {
-	destination := &models.Destination{
-		Name:          r.Name,
-		UniqueID:      r.UniqueID,
-		ConnectionURL: r.Connection.URL,
-		ConnectionCA:  string(r.Connection.CA),
-		Resources:     r.Resources,
-		Roles:         r.Roles,
-		Version:       r.Version,
-	}
-
-	err := access.CreateDestination(c, destination)
-	if err != nil {
-		return nil, fmt.Errorf("create destination: %w", err)
-	}
-
-	return destination.ToAPI(), nil
-}
-
-func (a *API) UpdateDestination(c *gin.Context, r *api.UpdateDestinationRequest) (*api.Destination, error) {
-	destination := &models.Destination{
-		Model: models.Model{
-			ID: r.ID,
-		},
-		Name:          r.Name,
-		UniqueID:      r.UniqueID,
-		ConnectionURL: r.Connection.URL,
-		ConnectionCA:  string(r.Connection.CA),
-		Resources:     r.Resources,
-		Roles:         r.Roles,
-		Version:       r.Version,
-	}
-
-	if err := access.SaveDestination(c, destination); err != nil {
-		return nil, fmt.Errorf("update destination: %w", err)
-	}
-
-	return destination.ToAPI(), nil
-}
-
-func (a *API) DeleteDestination(c *gin.Context, r *api.Resource) (*api.EmptyResponse, error) {
-	return nil, access.DeleteDestination(c, r.ID)
 }
 
 func (a *API) CreateToken(c *gin.Context, r *api.EmptyRequest) (*api.CreateTokenResponse, error) {
@@ -404,108 +87,6 @@ func (a *API) CreateAccessKey(c *gin.Context, r *api.CreateAccessKeyRequest) (*a
 		ExtensionDeadline: api.Time(accessKey.ExtensionDeadline),
 		AccessKey:         raw,
 	}, nil
-}
-
-func (a *API) ListGrants(c *gin.Context, r *api.ListGrantsRequest) (*api.ListResponse[api.Grant], error) {
-	var subject uid.PolymorphicID
-	p := models.RequestToPagination(r.PaginationRequest)
-	switch {
-	case r.User != 0:
-		subject = uid.NewIdentityPolymorphicID(r.User)
-	case r.Group != 0:
-		subject = uid.NewGroupPolymorphicID(r.Group)
-	}
-
-	grants, err := access.ListGrants(c, subject, r.Resource, r.Privilege, r.ShowInherited, &p)
-	if err != nil {
-		return nil, err
-	}
-
-	result := api.NewListResponse(grants, models.PaginationToResponse(p), func(grant models.Grant) api.Grant {
-		return *grant.ToAPI()
-	})
-
-	return result, nil
-}
-
-// TODO: remove after deprecation period
-func (a *API) deprecatedListUserGrants(c *gin.Context, r *api.Resource) (*api.ListResponse[api.Grant], error) {
-	return a.ListGrants(c, &api.ListGrantsRequest{User: r.ID})
-}
-
-// TODO: remove after deprecation period
-func (a *API) deprecatedListGroupGrants(c *gin.Context, r *api.Resource) (*api.ListResponse[api.Grant], error) {
-	return a.ListGrants(c, &api.ListGrantsRequest{Group: r.ID})
-}
-
-func (a *API) GetGrant(c *gin.Context, r *api.Resource) (*api.Grant, error) {
-	grant, err := access.GetGrant(c, r.ID)
-	if err != nil {
-		return nil, err
-	}
-
-	return grant.ToAPI(), nil
-}
-
-func (a *API) CreateGrant(c *gin.Context, r *api.CreateGrantRequest) (*api.CreateGrantResponse, error) {
-	var subject uid.PolymorphicID
-
-	switch {
-	case r.User != 0:
-		subject = uid.NewIdentityPolymorphicID(r.User)
-	case r.Group != 0:
-		subject = uid.NewGroupPolymorphicID(r.Group)
-	}
-
-	grant := &models.Grant{
-		Subject:   subject,
-		Resource:  r.Resource,
-		Privilege: r.Privilege,
-	}
-
-	err := access.CreateGrant(c, grant)
-	var ucerr data.UniqueConstraintError
-
-	if errors.As(err, &ucerr) {
-		grants, err := access.ListGrants(c, grant.Subject, grant.Resource, grant.Privilege, false, &models.Pagination{})
-
-		if err != nil {
-			return nil, err
-		}
-
-		if len(grants) == 0 {
-			return nil, fmt.Errorf("duplicate grant exists, but cannot be found")
-		}
-
-		return &api.CreateGrantResponse{Grant: grants[0].ToAPI()}, nil
-	}
-
-	if err != nil {
-		return nil, err
-	}
-
-	return &api.CreateGrantResponse{Grant: grant.ToAPI(), WasCreated: true}, nil
-
-}
-
-func (a *API) DeleteGrant(c *gin.Context, r *api.Resource) (*api.EmptyResponse, error) {
-	grant, err := access.GetGrant(c, r.ID)
-	if err != nil {
-		return nil, err
-	}
-
-	if grant.Resource == access.ResourceInfraAPI && grant.Privilege == models.InfraAdminRole {
-		infraAdminGrants, err := access.ListGrants(c, "", grant.Resource, grant.Privilege, false, &models.Pagination{})
-		if err != nil {
-			return nil, err
-		}
-
-		if len(infraAdminGrants) == 1 {
-			return nil, fmt.Errorf("%w: cannot remove the last infra admin", internal.ErrBadRequest)
-		}
-	}
-
-	return nil, access.DeleteGrant(c, r.ID)
 }
 
 func (a *API) SignupEnabled(c *gin.Context, _ *api.EmptyRequest) (*api.SignupEnabledResponse, error) {
@@ -648,34 +229,4 @@ func (a *API) providerClient(ctx context.Context, provider *models.Provider, red
 	}
 
 	return providers.NewOIDCClient(*provider, clientSecret, redirectURL), nil
-}
-
-// setProviderInfoFromServer checks information provided by an OIDC server
-func (a *API) setProviderInfoFromServer(c *gin.Context, provider *models.Provider) error {
-	// create a provider client to validate the server and get its info
-	oidc, err := a.providerClient(c, provider, "http://localhost:8301")
-	if err != nil {
-		return fmt.Errorf("%w: %s", internal.ErrBadRequest, err)
-	}
-
-	err = oidc.Validate(c)
-	if err != nil {
-		if errors.Is(err, providers.ErrValidation) {
-			return fmt.Errorf("%w: %s", internal.ErrBadRequest, err)
-		}
-		return err
-	}
-
-	authServerInfo, err := oidc.AuthServerInfo(c)
-	if err != nil {
-		if errors.Is(err, context.DeadlineExceeded) {
-			return fmt.Errorf("%w: %s", internal.ErrBadGateway, err)
-		}
-		return err
-	}
-
-	provider.AuthURL = authServerInfo.AuthURL
-	provider.Scopes = authServerInfo.ScopesSupported
-
-	return nil
 }
